Added files after refactoring
