+++ examples/hand_pose_estimation/processors/keypoints.py
@@ -1,64 +1,25 @@
 import numpy as np
 
-from backend.SE3 import rotation_from_axis_angles
-from backend.general import wrap_dictionary, merge_dictionaries
-from backend.handkeypoints import create_score_maps, extract_2D_keypoints
-from backend.handkeypoints import crop_image_from_coordinates, detect_keypoints
-from backend.handkeypoints import crop_image_using_mask, extract_hand_segment
-from backend.handkeypoints import extract_bounding_box, find_max_location
-from backend.handkeypoints import extract_dominant_hand_visibility
-from backend.handkeypoints import extract_dominant_keypoints2D, flip_right_hand
-from backend.handkeypoints import get_canonical_transformations
-from backend.handkeypoints import get_hand_side_and_keypooints
-from backend.handkeypoints import keypoint_to_root_frame
-from backend.handkeypoints import keypoints_to_palm_coordinates
-from backend.handkeypoints import normalize_keypoints
-from backend.handkeypoints import transform_cropped_keypoints
-from backend.handkeypoints import transform_visibility_mask
 from paz.abstract import Processor
 from paz.backend.image.tensorflow_image import resize
-from paz.backend.boxes import to_one_hot
+from ..backend.keypoints import create_score_maps, extract_2D_keypoints
+from ..backend.keypoints import crop_image_from_coordinates, detect_keypoints
+from ..backend.keypoints import crop_image_using_mask, extract_hand_segment
+from ..backend.keypoints import extract_bounding_box, find_max_location
+from ..backend.keypoints import extract_dominant_hand_visibility
+from ..backend.keypoints import extract_dominant_keypoints2D, flip_right_hand
+from ..backend.keypoints import get_hand_side_and_keypooints
+from ..backend.keypoints import normalize_keypoints
 
 
 class ExtractHandmask(Processor):
     """Extract Hand mask."""
+
     def __init__(self):
         super(ExtractHandmask, self).__init__()
 
     def call(self, segmentation_label):
         return extract_hand_segment(segmentation_label=segmentation_label)
-
-
-class ToOneHot(Processor):
-    """Extract Hand mask."""
-    def __init__(self, num_classes=2):
-        super(ToOneHot, self).__init__()
-        self.num_classes = num_classes
-
-    def call(self, class_indices):
-        return to_one_hot(class_indices, self.num_classes)
-
-
-class KeypointstoPalmFrame(Processor):
-    """Translate to Wrist Coordinates.
-    """
-
-    def __init__(self):
-        super(KeypointstoPalmFrame, self).__init__()
-
-    def call(self, keypoints):
-        return keypoints_to_palm_coordinates(keypoints=keypoints)
-
-
-class TransformVisibilityMask(Processor):
-    """Extract Visibility Mask.
-    """
-
-    def __init__(self):
-        super(TransformVisibilityMask, self).__init__()
-
-    def call(self, visibility_mask):
-        return transform_visibility_mask(visibility_mask)
 
 
 class ExtractHandSide(Processor):
@@ -81,27 +42,6 @@ class NormalizeKeypoints(Processor):
 
     def call(self, keypoints3D):
         return normalize_keypoints(keypoints3D)
-
-
-class TransformtoRelativeFrame(Processor):
-    """Transform to Relative Frame."""
-
-    def __init__(self):
-        super(TransformtoRelativeFrame, self).__init__()
-
-    def call(self, keypoints3D):
-        return np.squeeze(keypoint_to_root_frame(keypoints3D))
-
-
-class GetCanonicalTransformation(Processor):
-    """Extract Canonical Transformation matrix.
-        """
-
-    def __init__(self):
-        super(GetCanonicalTransformation, self).__init__()
-
-    def call(self, keypoints3D):
-        return get_canonical_transformations(keypoints3D)
 
 
 class FlipRightHand(Processor):
@@ -179,15 +119,6 @@ class Extract2DKeypoints(Processor):
         return extract_2D_keypoints(keypoint_visibility)
 
 
-class MatrixInverse(Processor):
-    """ Perform Pseudo Inverse of the matrix"""
-    def __init__(self):
-        super(MatrixInverse, self).__init__()
-
-    def call(self, matrix):
-        return np.linalg.pinv(matrix)
-
-
 class ExtractBoundingbox(Processor):
     """ Extract bounding box when provided with a binary mask"""
     def __init__(self):
@@ -223,34 +154,6 @@ class CropImage(Processor):
                                            scale)
 
 
-class RotationMatrixfromAxisAngles(Processor):
-    """ Get Rotation matrix from the axis angles"""
-    def __init__(self):
-        super(RotationMatrixfromAxisAngles, self).__init__()
-
-    def call(self, rotation_angles):
-        return rotation_from_axis_angles(rotation_angles)
-
-
-class CanonicaltoRelativeFrame(Processor):
-    """ Transform the keypoints from Canonical coordinates to chosen relative (
-    wrist or palm) coordinates """
-    def __init__(self, num_keypoints=21):
-        super(CanonicaltoRelativeFrame, self).__init__()
-        self.num_keypoints = num_keypoints
-
-    def call(self, canonical_coordinates, rotation_matrix, hand_side):
-        cond_right = np.equal(np.argmax(hand_side, 1), 1)
-        cond_right_all = np.tile(np.reshape(cond_right, [-1, 1, 1]),
-                                 [1, self.num_keypoints, 3])
-
-        coord_xyz_can_flip = flip_right_hand(canonical_coordinates,
-                                             cond_right_all)
-        # rotate view back
-        coord_xyz_rel_normed = np.matmul(coord_xyz_can_flip, rotation_matrix)
-        return coord_xyz_rel_normed
-
-
 class ExtractKeypoints(Processor):
     """ Extract keypoints when provided with a predicted scoremap"""
     def __init__(self):
@@ -270,30 +173,6 @@ class Resize_image(Processor):
         return resize(image, self.size)
 
 
-class WraptoDictionary(Processor):
-    """ Wrap the input values to a dictionary with already provided key
-    values """
-    def __init__(self, keys):
-        super(WraptoDictionary, self).__init__()
-        if not isinstance(keys, list):
-            keys = list(keys)
-        self.keys = keys
-
-    def call(self, values):
-        if not isinstance(values, list):
-            values = list(values)
-        return wrap_dictionary(self.keys, values)
-
-
-class MergeDictionaries(Processor):
-    """ Merge two dictionaries into one"""
-    def __init__(self):
-        super(MergeDictionaries, self).__init__()
-
-    def call(self, dicts):
-        return merge_dictionaries(dicts)
-
-
 class FindMaxLocation(Processor):
     """ Find the brightest point in the score map, which is represented as a
     keypoint"""
@@ -302,16 +181,4 @@ class FindMaxLocation(Processor):
 
     def call(self, scoremaps):
         keypoints_2D = find_max_location(scoremaps)
-        return keypoints_2D
-
-
-class TransformKeypoints(Processor):
-    """ Transform the keypoint from cropped image frame to original image
-    frame"""
-    def __init__(self):
-        super(TransformKeypoints, self).__init__()
-
-    def call(self, cropped_keypoints, centers, scale, crop_size):
-        keypoints_2D = transform_cropped_keypoints(cropped_keypoints, centers,
-                                                   scale, crop_size)
         return keypoints_2D
